+++ packages/asc-ui/src/utils/themeUtils.ts
@@ -339,12 +339,16 @@ export const showHide = () => ({ hideAt, showAt, theme }: ShowHideProps) => {
 // the backdrop, but aren't directly related
 export const showAboveBackDrop = (show?: boolean) => ({
   hasBackDrop,
+  zIndexOffset,
 }: {
   hasBackDrop?: boolean
+  zIndexOffset?: number
 }) =>
   hasBackDrop || show
     ? css`
-        z-index: ${BACKDROP_Z_INDEX + 1};
+        z-index: ${zIndexOffset
+          ? BACKDROP_Z_INDEX + zIndexOffset + 1
+          : BACKDROP_Z_INDEX + 1};
       `
     : ''